CRUD: Authorization and Authentication to FastAPI App

+++ Project3/routers/todos.py
@@ -70,6 +70,6 @@ async def delete_todo(user: user_dependency, db : db_dependency, delete_id: int 
     todo_model = db.query(Todos).filter(Todos.id == delete_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None: 
         raise HTTPException(status_code=404, detail="RECORD NOT FOUND")
-    db.query(Todos).filter(Todos.id == delete_id).delete()
+    db.query(Todos).filter(Todos.id == delete_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
     return {f"User {user.get('id')} DELETED TodoID{delete_id} SUCCESSFULLY"}
